server/requirements.txt
Django>=5.1,<7
django-ninja>=1.3
django-cors-headers>=4.3
channels>=4.0
daphne>=4.1
gunicorn>=21.2
whitenoise>=6.5
dj-database-url>=2.1
psycopg2-binary>=2.9
channels-redis>=4.2
discord.py>=2.3
slack-bolt>=1.18
python-dotenv>=1.0
google-auth>=2.28
requests>=2.31
